+++ vhdl_lang/src/completion/map_aspect.rs
@@ -187,6 +187,93 @@ mod tests {
     use crate::{list_completion_options, CompletionItem};
 
     #[test]
+    pub fn complete_component_instantiation_map() {
+        let mut input = LibraryBuilder::new();
+        let code = input.code(
+            "libname",
+            "\
+    entity my_ent is
+    end entity my_ent;
+
+    architecture arch of my_ent is
+        component comp is
+        generic (
+          A: natural := 5;
+          B: integer
+        );
+        port (
+          clk : in bit;
+          rst : in bit;
+          dout : out bit
+        );
+        end component comp;
+        signal clk, rst: bit;
+    begin
+        comp_inst: comp
+        generic map (
+            A => 2
+        )
+        port map (
+            clk => clk
+        );
+    end arch;
+            ",
+        );
+        let (root, _) = input.get_analyzed_root();
+        let cursor = code.s1("generic map (").pos().end();
+        let options = list_completion_options(&root, code.source(), cursor);
+        let ent = root
+            .search_reference(code.source(), code.s1("B").start())
+            .unwrap();
+        assert_eq!(options, vec![CompletionItem::Formal(ent)]);
+
+        let rst = root
+            .search_reference(code.source(), code.s1("rst").start())
+            .unwrap();
+
+        let dout = root
+            .search_reference(code.source(), code.s1("dout").start())
+            .unwrap();
+
+        let clk_signal = root
+            .search_reference(
+                code.source(),
+                code.s1("signal clk, rst: bit;").s1("clk").start(),
+            )
+            .unwrap();
+
+        let rst_signal = root
+            .search_reference(
+                code.source(),
+                code.s1("signal clk, rst: bit;").s1("rst").start(),
+            )
+            .unwrap();
+
+        let cursor = code.s1("port map (").pos().end();
+        let options = list_completion_options(&root, code.source(), cursor);
+        assert!(options.contains(&CompletionItem::Formal(rst)));
+        assert!(options.contains(&CompletionItem::Formal(dout)));
+        assert_eq!(options.len(), 2);
+        let cursor = code
+            .s1("port map (
+            clk =>")
+            .pos()
+            .end();
+        let options = list_completion_options(&root, code.source(), cursor);
+        assert!(options.contains(&CompletionItem::Simple(clk_signal)));
+        assert!(options.contains(&CompletionItem::Simple(rst_signal)));
+
+        let cursor = code
+            .s1("port map (
+            clk => c")
+            .pos()
+            .end();
+        let options = list_completion_options(&root, code.source(), cursor);
+        assert!(options.contains(&CompletionItem::Simple(clk_signal)));
+        assert!(options.contains(&CompletionItem::Simple(rst_signal)));
+    }
+
+    #[test]
     pub fn complete_in_generic_map() {
         let mut input = LibraryBuilder::new();
         let code = input.code(
