+++ vhdl_lang/src/syntax/recover.rs
@@ -1,3 +1,8 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at http://mozilla.org/MPL/2.0/.
+//
+// Copyright (c) 2024, Olof Kraigher olof.kraigher@gmail.com
 /// Module for robust parsing
 use crate::syntax::parser::ParsingContext;
 use crate::syntax::Kind::{Colon, SemiColon};
